Sectors are now marked and unmarked with switch ID, instead of with player number
I realized that if the same player activated multiple traps that affected
the same area, the first trap that turns off would unmark the overlapping
sectors for all of that player's traps, which is not desired.

This also allows for a blanket sector unmarking by switch ID - good for
cleanup.